Update types of legacy tracker model to match to domain one (#245)
* `score` to Double

* `tracker_id` to Long

* `last_chapter_read` to Double

* `total_chapters` to Long

* `status` to Long

(cherry picked from commit 05efc4ebeb197cabd2bc494d36fbb27fea923292)

+++ app/src/main/java/eu/kanade/tachiyomi/data/track/mdlist/MdList.kt
@@ -40,18 +40,18 @@ class MdList(id: Long) : BaseTracker(id, "MDList") {
         return Color.rgb(43, 48, 53)
     }
 
-    override fun getStatusList(): List<Int> {
+    override fun getStatusList(): List<Long> {
         return FollowStatus.entries.map { it.int }
     }
 
-    override fun getStatus(status: Int): StringResource? = when (status) {
-        0 -> SYMR.strings.md_follows_unfollowed
-        1 -> MR.strings.reading
-        2 -> MR.strings.completed
-        3 -> MR.strings.on_hold
-        4 -> MR.strings.plan_to_read
-        5 -> MR.strings.dropped
-        6 -> MR.strings.repeating
+    override fun getStatus(status: Long): StringResource? = when (status) {
+        0L -> SYMR.strings.md_follows_unfollowed
+        1L -> MR.strings.reading
+        2L -> MR.strings.completed
+        3L -> MR.strings.on_hold
+        4L -> MR.strings.plan_to_read
+        5L -> MR.strings.dropped
+        6L -> MR.strings.repeating
         else -> null
     }
 
@@ -103,11 +103,11 @@ class MdList(id: Long) : BaseTracker(id, "MDList") {
         }
     }
 
-    override fun getCompletionStatus(): Int = FollowStatus.COMPLETED.int
+    override fun getCompletionStatus(): Long = FollowStatus.COMPLETED.int
 
-    override fun getReadingStatus(): Int = FollowStatus.READING.int
+    override fun getReadingStatus(): Long = FollowStatus.READING.int
 
-    override fun getRereadingStatus(): Int = FollowStatus.RE_READING.int
+    override fun getRereadingStatus(): Long = FollowStatus.RE_READING.int
 
     override suspend fun bind(track: Track, hasReadChapters: Boolean): Track = update(
         refresh(track).also {
